login_app/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
PyJWT==2.8.0
requests==2.31.0
